raw_data/requirements.txt
asyncpg==0.29.0
fastapi==0.110.0
httpx==0.27.0
pandas==2.2.1
pandera==0.18.0
pydantic-settings==2.2.1
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-xdist==3.5.0
rich==13.7.1
uvicorn==0.27.1
sqlmodel == 0.0.16
